return NoProofFound when hitting cache in FormSubRequest
The method forgot to check if the proof was valid when retrieving it
from the cache. If it is *not* valid, then we return the NoProofFound
error like in FormQueryRequest.

This caused a situation where if a client attempted to re-subscribe
after first failing to subscribe due to a lack of a proof, the client
would successfully get the server to allocate subscription resources,
even though the upstream connection wasn't set